RealTimeOpenAI-Basic/requirements.txt
openai>=1.40.0
pyaudio>=0.2.11
speech_recognition>=3.10.0
colorama>=0.4.6
python-dotenv>=1.0.0
numpy>=1.24.0
pydub>=0.25.1
websockets>=12.0
webrtcvad>=2.0.10
aiohttp>=3.8.0